use "run()" to handle general cmds with error msg
For the commands non-specified general error message, we now use the new function "run()" to handle these commands,
the error message will now contain the running commands to help debug!

PS, in line 133, I'm not sure how to handle the one-time variable for internal function, so just separate the reindex command for now.